Work on level generator.
Slight refactor needed.
The way the generation works (figuring out the connection before we
actually generate the next room) is a bit weird, and causing some problems,
as it makes it harder to record the connections.
I think I should rework it to take the current room, create the new room,
and then create the connection from there, with a function on ARoom like:
ConnectRoom(int Index, ARoom* Other, int OtherIndex).